zenq/requirements_dev.txt
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Click==7.1.2
pytest==6.2.4
black==21.7b0
psycopg2-binary==2.9.5
